Factory pattern for the view models
