+++ RevenueCatUI/Templates/Components/Packages/Package/PackageComponentViewModel.swift
@@ -19,34 +19,23 @@ import RevenueCat
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
 class PackageComponentViewModel {
 
-    private let localizedStrings: PaywallComponent.LocalizationDictionary
-    private let component: PaywallComponent.PackageComponent
-    private let offering: Offering
-
     let isSelectedByDefault: Bool
     let package: Package?
     let stackViewModel: StackComponentViewModel
 
-    init(packageValidator: PackageValidator,
-         localizedStrings: PaywallComponent.LocalizationDictionary,
-         component: PaywallComponent.PackageComponent,
-         offering: Offering) throws {
-        self.localizedStrings = localizedStrings
-        self.component = component
-        self.offering = offering
-
+    init(
+        localizedStrings: PaywallComponent.LocalizationDictionary,
+        component: PaywallComponent.PackageComponent,
+        offering: Offering,
+        stackViewModel: StackComponentViewModel
+    ) {
         self.isSelectedByDefault = component.isSelectedByDefault
         self.package = offering.package(identifier: component.packageID)
         if package == nil {
             Logger.warning(Strings.paywall_could_not_find_package(component.packageID))
         }
 
-        self.stackViewModel = try StackComponentViewModel(
-            packageValidator: packageValidator,
-            component: component.stack,
-            localizedStrings: localizedStrings,
-            offering: offering
-        )
+        self.stackViewModel = stackViewModel
     }
 
 }
