+++ RevenueCatUI/Templates/Components/Packages/PurchaseButton/PurchaseButtonComponentViewModel.swift
@@ -20,26 +20,12 @@ import SwiftUI
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
 class PurchaseButtonComponentViewModel {
 
-    private let localizedStrings: PaywallComponent.LocalizationDictionary
-    private let component: PaywallComponent.PurchaseButtonComponent
-
     let stackViewModel: StackComponentViewModel
 
     init(
-        packageValidator: PackageValidator,
-        localizedStrings: PaywallComponent.LocalizationDictionary,
-        component: PaywallComponent.PurchaseButtonComponent,
-        offering: Offering
-    ) throws {
-        self.localizedStrings = localizedStrings
-        self.component = component
-
-        self.stackViewModel = try StackComponentViewModel(
-            packageValidator: packageValidator,
-            component: component.stack,
-            localizedStrings: localizedStrings,
-            offering: offering
-        )
+        stackViewModel: StackComponentViewModel
+    ) {
+        self.stackViewModel = stackViewModel
     }
 
 }
